fastapi_demo/pyproject.toml
[project]
name = "sapl-fastapi-demo"
version = "1.0.0"
description = "FastAPI demo application for SAPL Python PEP"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "sapl-fastapi>=4.0.0a1",
    "pyjwt[crypto]>=2.8",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "structlog>=24.1",
    "sse-starlette>=2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
